Refactor out `lovatAPI.getScoutReportsByScouter()`
