+++ lib/pages/scouters.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:frc_8033_scouting_shared/frc_8033_scouting_shared.dart';
 import 'package:scouting_dashboard_app/datatypes.dart';
 import 'package:scouting_dashboard_app/reusable/friendly_error_view.dart';
+import 'package:scouting_dashboard_app/reusable/lovat_api/get_scout_reports_by_scouter.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/get_scouter_overviews.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/lovat_api.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/scouter.dart';
